requirements.txt
# requirements.txt for CLV Engine System

# Core Data Libraries
pandas>=1.3.0
numpy>=1.21.0

# Modeling and Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.0.0

# Visualization (only for clv_grid_search_autotune.py)
matplotlib>=3.4.0
seaborn>=0.11.0

# Config Management
PyYAML>=5.4.0

# Statistical Tests
scipy>=1.7.0
